requirements.txt
numpy==2.0.1
onnxruntime==1.19.0
opencv-python==4.10.0.84
pillow==10.2.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
